Fixed a small warning
git-svn-id: file:///D/data/svnrepos/uimager/trunk@2550 15f0fa91-3113-114a-a8b1-9261d5763783

+++ LicenseHelper.cpp
@@ -79,7 +79,7 @@ int CDonorEmailValidateThread::DonorEmailValidate()
 			p = ::GetURL(sURL, Size, FALSE, TRUE, TRUE, NULL);		// 2. then try with proxy
 		if (p)
 		{
-			CString s(::FromUTF8(p, Size));
+			CString s(::FromUTF8(p, (int)Size));
 			free(p);
 			if (s.Find(_T("OK")) >= 0)
 			{
